Fix crash if plugin was null

+++ src/controllers/api/v1/plugins.js
@@ -83,10 +83,13 @@ api_plugins.installPlugin = function(req, res) {
 api_plugins.removePlugin = function(req, res) {
     var packageid = req.params.packageid;
 
-    request.get(pluginServerUrl + '/api/plugin/package/' + packageid, function(err, response) {
+    request.get(pluginServerUrl + '/api/plugin/package/' + packageid, function(err, response, body) {
         if (err) return res.status(400).json({success: false, error: err});
 
-        var plugin = JSON.parse(response.body).plugin;
+        var plugin = JSON.parse(body).plugin;
+
+        if (plugin == null)
+            return res.json({success: false, error: 'Invalid Plugin'});
 
         rimraf(path.join(pluginPath, plugin.name.toLowerCase()), function(err) {
             if (err) return res.json({success: false, error: 'Unable to remove plugin directory.'});
